Add patch to add GitHub action to release any changes to staging.
This adds a GitHub action as a patch, which releases all changes automatically to the staging environment for the greeter service, by building the image, checking if the sha sum updates the deployment yaml, and then opening a pull request to make the release.

The patch needs to be applied with `git apply .github/workflows/add-release-actions-workflow.patch` and then committed/pushed to activate it.